+++ src/Resources/app/administration/src/main.js
@@ -32,7 +32,10 @@ import './module/sw-cms/elements/button';
 import './module/sw-cms/blocks/ultimate-cms-tools/button';
 import './module/sw-cms/elements/icon-list';
 import './module/sw-cms/blocks/ultimate-cms-tools/icon-list';
+import './module/sw-cms/elements/flexible-image-text';
+import './module/sw-cms/blocks/ultimate-cms-tools/flexible-image-text';
 import deDE from './module/sw-cms/snippet/de-DE.json';
+
 import enGB from './module/sw-cms/snippet/en-GB.json';
 import nlNL from './module/sw-cms/snippet/nl-NL.json';