+++ src/Resources/app/administration/src/module/sw-cms/blocks/ultimate-cms-tools/flexible-image-text/component/index.js
@@ -1,0 +1,8 @@
+import template from './sw-cms-block-flexible-image-text.html.twig';
+import './sw-cms-block-flexible-image-text.scss';
+
+const { Component } = Shopware;
+
+Component.register('sw-cms-block-flexible-image-text', {
+    template,
+});