+++ src/Resources/app/storefront/src/main.js
@@ -5,6 +5,7 @@ import CustomCarouselPlugin from './plugin/custom-carousel.plugin';
 import CustomProductCarouselPlugin from './plugin/custom-product-carousel.plugin';
 import StatisticsCounterPlugin from './plugin/statistics-counter.plugin';
 import IconListAnimationPlugin from './plugin/icon-list-animation.plugin';
+import FlexibleImageTextAnimationPlugin from './plugin/flexible-image-text-animation.plugin';
 
 window.PluginManager.register('SubcategoryCarousel', SubcategoryCarouselPlugin, '[data-subcategory-carousel]');
 window.PluginManager.register('ManufacturerCarousel', ManufacturerCarouselPlugin, '[data-manufacturer-carousel]');
@@ -13,5 +14,7 @@ window.PluginManager.register('CustomCarousel', CustomCarouselPlugin, '[data-cus
 window.PluginManager.register('CustomProductCarousel', CustomProductCarouselPlugin, '[data-custom-product-carousel]');
 window.PluginManager.register('StatisticsCounter', StatisticsCounterPlugin, '[data-statistics-counter]');
 window.PluginManager.register('IconListAnimation', IconListAnimationPlugin, '[data-icon-list-animation]');
+window.PluginManager.register('FlexibleImageTextAnimation', FlexibleImageTextAnimationPlugin, '[data-flexible-image-text-animation]');
+